Refactor create-slots rename create-slots -> create-slot
